Перенес фильтр директорий для хендлера в utils
