+++ utils/utils.py
@@ -1,5 +1,8 @@
+from aiogram.filters import BaseFilter
+from aiogram import types
+
 from keyboards.user_keyboards import system_buttons
-from file_managment.ya_file_manager import path_list
+from file_managment.ya_file_manager import path_list, YaDisk
 
 
 def move_to_folders_on_disk(folder) -> None:
@@ -13,3 +16,12 @@ def move_to_folders_on_disk(folder) -> None:
     if folder not in ['', *system_buttons.values()]:
         path_list.append(folder)
 
+
+class FolderFilter(BaseFilter):
+    """Хендлер кнопок для предстоящего перехода по директориям"""
+    async def __call__(self, message: types.Message) -> bool:
+        result_handler_list = await YaDisk.get_directories_in(path_list) + list(system_buttons.values())
+        if message.text in result_handler_list:
+            return True
+        else:
+            return False
